added logo to README
